requirements.txt
aws-cdk-lib==2.93.0
tonstructs>=10.0.0,<11.0.0
